initramfs: Add recipe for tee-supplicant hook
This adds the tee-supplicant hook so that the tee supplicant daemon is
started at the initrd stage.

the tee-supplicant daemon is used to provide service to trust
applications running in optee, for example to provide RPMB access
service for StMM or fTPM TAs.

By running tee-supplicant at initrd stage, disk encryption based on fTPM
is possible.

Signed-off-by: Baocheng Su <baocheng.su@siemens.com>